requirements.txt
requests==2.32.3
httpx==0.23.1
beautifulsoup4==4.13.3
lxml==5.3.1
lxml_html_clean==0.4.1
newspaper3k==0.2.8
googlesearch-python==1.3.0
psycopg2-binary==2.9.10